Expand pandas merging notebook
A couple examples using string methods were added to the merging
notebook, as they have a lot of utility when working with data from
multiple sources and trying to merge that data together.

In addition, in order for the students to have all of the tools
necessary to complete the coding task for the week, a couple of
examples of working with numeric data were added, to be further
expanded upon in week three when we get into exploratory data analysis.